#1799 mv functionality into structure_reference and generalise
