+++ src/psyclone/psyir/nodes/array_of_structures_reference.py
@@ -39,7 +39,6 @@ node. '''
 
 # Circular import if only '...nodes' is used:
 from psyclone.psyir import symbols
-from psyclone.psyir.nodes.array_mixin import ArrayMixin
 from psyclone.psyir.nodes.array_of_structures_mixin import (
     ArrayOfStructuresMixin)
 from psyclone.psyir.nodes.structure_reference import StructureReference
@@ -126,38 +125,6 @@ class ArrayOfStructuresReference(ArrayOfStructuresMixin, StructureReference):
         for child in indices:
             ref.addchild(child)
         return ref
-
-    @property
-    def datatype(self):
-        '''
-        :returns: the datatype of this reference.
-        :rtype: :py:class:`psyclone.psyir.symbols.DataType`
-        '''
-
-        if isinstance(self.symbol.datatype, symbols.DeferredType):
-            return symbols.DeferredType()
-        if not isinstance(self.symbol.datatype.intrinsic,
-                          symbols.DataTypeSymbol):
-            return symbols.DeferredType()
-        dtype = self.symbol.datatype.intrinsic.datatype
-        if isinstance(dtype, symbols.DeferredType):
-            return symbols.DeferredType()
-
-        cursor = self
-        cursor_type = dtype
-        shape = cursor.shape
-        try:
-            while cursor.member:
-                cursor = cursor.member
-                cursor_type = cursor_type.components[cursor.name]
-                if isinstance(cursor, ArrayMixin):
-                    shape.extend(cursor.shape)
-        except AttributeError:
-            pass
-
-        if shape:
-            return symbols.ArrayType(cursor_type.datatype, shape)
-        return cursor_type.datatype
 
 
 # For AutoAPI documentation generation
