+++ src/psyclone/psyir/nodes/structure_reference.py
@@ -1,7 +1,7 @@
 # -----------------------------------------------------------------------------
 # BSD 3-Clause License
 #
-# Copyright (c) 2020-2021, Science and Technology Facilities Council.
+# Copyright (c) 2020-2022, Science and Technology Facilities Council.
 # All rights reserved.
 #
 # Redistribution and use in source and binary forms, with or without
@@ -37,18 +37,16 @@
 
 ''' This module contains the implementation of the StructureReference node. '''
 
-from __future__ import absolute_import
-import six
-
 from psyclone.core import Signature
 from psyclone.psyir.nodes.reference import Reference
 from psyclone.psyir.nodes.member import Member
 from psyclone.psyir.nodes.array_member import ArrayMember
+from psyclone.psyir.nodes.array_mixin import ArrayMixin
 from psyclone.psyir.nodes.array_of_structures_member import \
     ArrayOfStructuresMember
 from psyclone.psyir.nodes.structure_member import StructureMember
-from psyclone.psyir.symbols import DataSymbol, DataTypeSymbol, StructureType, \
-    DeferredType, UnknownType
+from psyclone.psyir.symbols import (DataSymbol, DataTypeSymbol, StructureType,
+                                    ArrayType, DeferredType, UnknownType)
 from psyclone.errors import InternalError
 
 
@@ -168,7 +166,7 @@ class StructureReference(Reference):
         if isinstance(members[-1], tuple):
             # An access to one or more array elements
             subref = ArrayMember.create(members[-1][0], members[-1][1])
-        elif isinstance(members[-1], six.string_types):
+        elif isinstance(members[-1], str):
             # A member access
             subref = Member(members[-1])
         else:
@@ -188,7 +186,7 @@ class StructureReference(Reference):
                 # This is an array access so we have an ArrayOfStructuresMember
                 subref = ArrayOfStructuresMember.create(
                     component[0], component[1], subref)
-            elif isinstance(component, six.string_types):
+            elif isinstance(component, str):
                 # No array access so just a StructureMember
                 subref = StructureMember.create(component, subref)
             else:
@@ -239,6 +237,41 @@ class StructureReference(Reference):
         sub_sig, indices = self.children[0].get_signature_and_indices()
         # Combine signature and indices
         return (Signature(my_sig, sub_sig), my_index + indices)
+
+    @property
+    def datatype(self):
+        '''
+        :returns: the datatype of this reference.
+        :rtype: :py:class:`psyclone.psyir.symbols.DataType`
+        '''
+        if isinstance(self.symbol.datatype, DeferredType):
+            return DeferredType()
+        if not isinstance(self.symbol.datatype.intrinsic,
+                          DataTypeSymbol):
+            return DeferredType()
+        dtype = self.symbol.datatype.intrinsic.datatype
+        if isinstance(dtype, DeferredType):
+            return dtype
+
+        cursor = self
+        cursor_type = dtype
+        if isinstance(cursor, ArrayMixin):
+            shape = cursor.shape
+        else:
+            shape = []
+
+        try:
+            while cursor.member:
+                cursor = cursor.member
+                cursor_type = cursor_type.components[cursor.name]
+                if isinstance(cursor, ArrayMixin):
+                    shape.extend(cursor.shape)
+        except AttributeError:
+            pass
+
+        if shape:
+            return ArrayType(cursor_type.datatype, shape)
+        return cursor_type.datatype
 
 
 # For AutoAPI documentation generation
